+++ src/Reducer/Appending.php
