+++ src/Reducer/Mapping.php
