+++ src/Reducer/SingleResult.php
